+++ src/main/java/org/threeten/bp/.walkingkooka-j2cl-maven-plugin-ignored-files.txt
@@ -1,0 +1,2 @@
+Ser.*
+
