+++ src/test/java/org/threeten/bp/TestZonedDateTime.java
@@ -177,16 +177,16 @@ public class TestZonedDateTime extends AbstractDateTimeTest {
     }
 
     //-----------------------------------------------------------------------
-    @Test
-    public void test_serialization() throws ClassNotFoundException, IOException {
-        assertSerializable(TEST_DATE_TIME);
-    }
-
-    @Test
-    public void test_serialization_format() throws ClassNotFoundException, IOException {
-        ZonedDateTime zdt = LocalDateTime.of(2012, 9, 16, 22, 17, 59, 470 * 1000000).atZone(ZoneId.of("Europe/London"));
-        assertEqualsSerialisedForm(zdt);
-    }
+//    @Test
+//    public void test_serialization() throws ClassNotFoundException, IOException {
+//        assertSerializable(TEST_DATE_TIME);
+//    }
+//
+//    @Test
+//    public void test_serialization_format() throws ClassNotFoundException, IOException {
+//        ZonedDateTime zdt = LocalDateTime.of(2012, 9, 16, 22, 17, 59, 470 * 1000000).atZone(ZoneId.of("Europe/London"));
+//        assertEqualsSerialisedForm(zdt);
+//    }
 
     //-----------------------------------------------------------------------
     // now()
